+++ src/test/java/org/opendevstack/provision/services/JenkinsPipelineAdapterTest.java
@@ -29,7 +29,6 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 import org.junit.runner.RunWith;
-import org.mockito.InjectMocks;
 import org.mockito.Mockito;
 import org.mockito.MockitoAnnotations;
 import org.opendevstack.provision.SpringBoot;
@@ -42,6 +41,8 @@ import org.opendevstack.provision.model.jenkins.Option;
 import org.opendevstack.provision.model.webhookproxy.CreateProjectResponse;
 import org.opendevstack.provision.util.CreateProjectResponseUtil;
 import org.opendevstack.provision.util.ValueCaptor;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Value;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
 import org.springframework.http.HttpMethod;
@@ -58,7 +59,10 @@ public class JenkinsPipelineAdapterTest extends AbstractBaseServiceAdapterTest {
 
   private static final String PROJECT_KEY = "123key";
 
-  @InjectMocks JenkinsPipelineAdapter jenkinsPipelineAdapter;
+  @Autowired private JenkinsPipelineAdapter jenkinsPipelineAdapter;
+
+  @Value("${jenkinspipeline.create-project.default-project-groups}")
+  private String defaultEntitlementGroups;
 
   private final String JOB_1_NAME = "be-java-springboot";
   private final String JOB_1_REPO = "gitRepoName.git";
@@ -87,7 +91,6 @@ public class JenkinsPipelineAdapterTest extends AbstractBaseServiceAdapterTest {
     jenkinsPipelineAdapter.bitbucketOdsProject = "opendevstack";
     jenkinsPipelineAdapter.odsImageTag = "latest";
     jenkinsPipelineAdapter.odsGitRef = "master";
-    jenkinsPipelineAdapter.clusterAdminsEntitlementGroups = "ADMINGROUP=MY-ADMIN-GROUP";
     jenkinsPipelineAdapter.init();
     super.beforeTest();
   }
@@ -407,7 +410,7 @@ public class JenkinsPipelineAdapterTest extends AbstractBaseServiceAdapterTest {
 
     String groups = actualBody.getOptionValue("PROJECT_GROUPS");
     assertNotNull(groups);
-    Assertions.assertThat(groups).contains(jenkinsPipelineAdapter.clusterAdminsEntitlementGroups);
+    Assertions.assertThat(groups).contains(defaultEntitlementGroups);
     Assertions.assertThat(groups).contains("ADMINGROUP=" + projectData.projectAdminGroup);
     Assertions.assertThat(groups).contains("USERGROUP=" + projectData.projectUserGroup);
     Assertions.assertThat(groups).contains("READONLYGROUP=" + projectData.projectReadonlyGroup);